transcription/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
decorator==5.1.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
msgpack==1.0.8
numba==0.59.1
numpy==1.26.4
openai==1.14.2
packaging==24.0
platformdirs==4.2.0
pooch==1.8.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
